+++ chain/rust/src/utils.rs
@@ -196,7 +196,7 @@ enum BigIntEncoding {
 
 #[derive(Clone, Debug, Derivative)]
 #[derivative(Eq, PartialEq, Ord, PartialOrd, Hash)]
-pub struct BigInt {
+pub struct BigInteger {
     num: num_bigint::BigInt,
     #[derivative(
         PartialEq = "ignore",
@@ -207,7 +207,7 @@ pub struct BigInt {
     encoding: Option<BigIntEncoding>,
 }
 
-impl serde::Serialize for BigInt {
+impl serde::Serialize for BigInteger {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: serde::Serializer,
@@ -216,14 +216,14 @@ impl serde::Serialize for BigInt {
     }
 }
 
-impl<'de> serde::de::Deserialize<'de> for BigInt {
+impl<'de> serde::de::Deserialize<'de> for BigInteger {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
         D: serde::de::Deserializer<'de>,
     {
         use std::str::FromStr;
         let s = <String as serde::de::Deserialize>::deserialize(deserializer)?;
-        BigInt::from_str(&s).map_err(|_e| {
+        BigInteger::from_str(&s).map_err(|_e| {
             serde::de::Error::invalid_value(
                 serde::de::Unexpected::Str(&s),
                 &"string rep of a big int",
@@ -232,9 +232,9 @@ impl<'de> serde::de::Deserialize<'de> for BigInt {
     }
 }
 
-impl schemars::JsonSchema for BigInt {
+impl schemars::JsonSchema for BigInteger {
     fn schema_name() -> String {
-        String::from("BigInt")
+        String::from("BigInteger")
     }
     fn json_schema(gen: &mut schemars::gen::SchemaGenerator) -> schemars::schema::Schema {
         String::json_schema(gen)
@@ -244,13 +244,13 @@ impl schemars::JsonSchema for BigInt {
     }
 }
 
-impl std::fmt::Display for BigInt {
+impl std::fmt::Display for BigInteger {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         self.num.fmt(f)
     }
 }
 
-impl std::str::FromStr for BigInt {
+impl std::str::FromStr for BigInteger {
     type Err = num_bigint::ParseBigIntError;
     fn from_str(string: &str) -> Result<Self, Self::Err> {
         num_bigint::BigInt::from_str(string).map(|num| Self {
@@ -260,7 +260,7 @@ impl std::str::FromStr for BigInt {
     }
 }
 
-impl BigInt {
+impl BigInteger {
     // can't be a trait due to being in other crate
     pub fn from_int(x: &Int) -> Self {
         Self {
@@ -325,7 +325,7 @@ impl BigInt {
     }
 }
 
-impl Serialize for BigInt {
+impl Serialize for BigInteger {
     fn serialize<'se, W: Write>(
         &self,
         serializer: &'se mut Serializer<W>,
@@ -413,7 +413,7 @@ impl Serialize for BigInt {
     }
 }
 
-impl Deserialize for BigInt {
+impl Deserialize for BigInteger {
     fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
         (|| -> Result<_, DeserializeError> {
             match raw.cbor_type()? {
@@ -468,11 +468,11 @@ impl Deserialize for BigInt {
                 _ => Err(DeserializeFailure::NoVariantMatched.into()),
             }
         })()
-        .map_err(|e| e.annotate("BigInt"))
+        .map_err(|e| e.annotate("BigInteger"))
     }
 }
 
-impl<T> std::convert::From<T> for BigInt
+impl<T> std::convert::From<T> for BigInteger
 where
     T: std::convert::Into<num_bigint::BigInt>,
 {
@@ -553,7 +553,7 @@ mod tests {
     #[test]
     fn bigint_uint_min() {
         let bytes = [0x00];
-        let x = BigInt::from_cbor_bytes(&bytes).unwrap();
+        let x = BigInteger::from_cbor_bytes(&bytes).unwrap();
         assert_eq!(bytes, x.to_cbor_bytes().as_slice());
         assert_eq!(x.as_u64(), Some(u64::MIN));
         assert_eq!(x.as_int().unwrap().to_string(), x.to_string());
@@ -563,7 +563,7 @@ mod tests {
     #[test]
     fn bigint_uint_max() {
         let bytes = [0x1B, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF];
-        let x = BigInt::from_cbor_bytes(&bytes).unwrap();
+        let x = BigInteger::from_cbor_bytes(&bytes).unwrap();
         assert_eq!(bytes, x.to_cbor_bytes().as_slice());
         assert_eq!(x.as_u64(), Some(u64::MAX));
         assert_eq!(x.as_int().unwrap().to_string(), x.to_string());
@@ -575,7 +575,7 @@ mod tests {
         let bytes = [
             0xC2, 0x49, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
         ];
-        let x = BigInt::from_cbor_bytes(&bytes).unwrap();
+        let x = BigInteger::from_cbor_bytes(&bytes).unwrap();
         assert_eq!(bytes, x.to_cbor_bytes().as_slice());
         assert_eq!(x.as_int(), None);
         assert_eq!(x.to_string(), "18446744073709551616");
@@ -584,7 +584,7 @@ mod tests {
     #[test]
     fn bigint_nint_min() {
         let bytes = [0x3B, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF];
-        let x = BigInt::from_cbor_bytes(&bytes).unwrap();
+        let x = BigInteger::from_cbor_bytes(&bytes).unwrap();
         assert_eq!(bytes, x.to_cbor_bytes().as_slice());
         assert_eq!(
             Into::<i128>::into(&x.as_int().unwrap()),
@@ -597,7 +597,7 @@ mod tests {
     #[test]
     fn bigint_nint_max() {
         let bytes = [0x20];
-        let x = BigInt::from_cbor_bytes(&bytes).unwrap();
+        let x = BigInteger::from_cbor_bytes(&bytes).unwrap();
         assert_eq!(bytes, x.to_cbor_bytes().as_slice());
         assert_eq!(x.as_u64(), None);
         assert_eq!(x.as_int().unwrap().to_string(), x.to_string());
@@ -609,7 +609,7 @@ mod tests {
         let bytes = [
             0xC3, 0x49, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
         ];
-        let x = BigInt::from_cbor_bytes(&bytes).unwrap();
+        let x = BigInteger::from_cbor_bytes(&bytes).unwrap();
         assert_eq!(bytes, x.to_cbor_bytes().as_slice());
         assert_eq!(x.as_int(), None);
         assert_eq!(x.to_string(), "-18446744073709551617");
